+++ packages/eslint/src/config.ts
@@ -13,13 +13,6 @@ type Config = ReturnType<typeof antfu>
 /**
  * @param antfuOptions Configures for antfu's config.
  * @param userFlatConfigs From the second arguments they are ESLint Flat Configs, you can have multiple configs.
- *
- * 前提：用户预设优于内置预设，若有相同的 key，则进行覆盖。
- * 1. 排除 rules 的前提下，合并 baseOptions 与 userOptions，得到 mergedOptions
- * 2. 根据 mergedOptions 取预设规则 presetRules，不对用户 rules 作错误兼容
- * 3. 将 presetRules 与 userRules 合并，得到 mergedRules
- * 4. 将 mergedOptions 与 { rules: mergedRules } 合并，得到 resolvedOptions
- * 5. 将 resolvedOptions 作为 antfu 第一个参数
  */
 export function defineConfig(
   antfuOptions?: AntfuOptions,